refactor(#475): extract shared did:key codec to de-duplicate native/wasm
Review finding [CLEANUP] iroh_peer.rs:86 — the did:key encode/decode logic
was duplicated between native `did_web.rs` and wasm32 `iroh_peer.rs` with
divergent behavior (no shared MULTICODEC_ED25519_PUB constant, no fragment
stripping on the wasm side).

`did_web` is native-only (cfg(not(wasm32))) and `iroh_peer` is wasm32-only, so
they cannot reference each other. Extract the Ed25519 did:key codec
(MULTICODEC_ED25519_PUB, decode_ed25519_multikey, did_key_to_ed25519,
ed25519_to_did_key) into a new cross-target `did_key` module compiled on all
targets, the single source of truth both paths share. `did_web` re-exports the
public fns so existing `crate::did_web::*` callers (admission gate, federation
tests) are unaffected.
